depfiles.c: switch on dlen
It occurred to me that the code is way too clever; when I die,
it will be hard to update the code (e.g. to add another dir),
and Alexey Tourbin will be cursed by his feeble-minded survivors
for writing such clever code.

So reasonable requirements to which I cede might be as follows:
1) The directories must be enumerated in an easy-to-grasp way.
2) The directory names must be spelled in full, not in 4-byte parts.

It turns out if you just switch on dlen, you can easily fulfill the
requirements and also get 3x faster code, as per valgrind instructions!