depfiles.c: implemented inplace stripFileList!
The rpmlib API is heavy and wasteful, it does things like sorting
the header entries which in our case are already sorted (so one can
fecklessly behold glibc's msort_with_tmp at the very top of valgrind
output).  But of course, I do not intend to reimplement the whole API.
Therefore, the plan is to use the API only when loading headers
from .rpm packages.  This will also play nice with rpmhdrcache.so,
which wraps some parts of the API.  On the other hand, when a blob
comes from the previous output, it can and should be handled without
the API.  Only two ops are required, I believe: findDepFiles and
stripFileList.  Both are now implemented.  In other words, once
a blob has been cooked with the API, it is afterwards susceptible
only to a limited set of ops which deal with filenames.  Generally
there is no point in scanning or changing other parts of the blob.