depfiles.c: optimize findDepFilesH1, tweak addDepFile
There were 4 addDepFile() calls in the code, and now there are 5.  In
findDepFilesB1, it was inlined 2 times in the same loop.  Unfortunately,
I wasn't able to make gcc produce two addDepFile instances, one for
hasLen=true and the other for hasLen=false.  (Since there was only one
branch that depended on hasLen, gcc might be right after all.)  Then
I noticed that the only call with hasLen=false is in findDepFilesH1.
As explained in the code, the length can actually be recovered for all
but the very last RequireName/ProvideName.  Since the length is almost
always known, I strip the last arg and make the function non-inline.
This opens the possibility for another optimization: the test for
a close paren can be done first, before the more expensive test for
bindir.  gcc now does the right thing by inlining the first test and
factoring addDepFile.part.2.