md5cache.c: added md5nocache()
After an anxious night's sleep, I'm no longer sure that the cache is
such a good idea.  It's slow on the first run, 2-3 times slower than just
reading srpms, because it does msync() for each record added.  After the
first run, most of these md5 sums will never be retrieved again, because
of another mechanism kicking in, namely the reuse of a previous output
as a cache.